Add statisticts for 2 game

+++ backend/app/db/models/false_state_stats.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from sqlalchemy import Column, ForeignKey
-from sqlalchemy.dialects.postgresql import BOOLEAN, INTEGER, TIMESTAMP
+from sqlalchemy.dialects.postgresql import BOOLEAN, INTEGER, TIMESTAMP, BIGINT
 from sqlalchemy.orm import relationship
 
 from backend.app.db import DeclarativeBase
@@ -11,7 +11,7 @@ class FalseStateStats(DeclarativeBase):
     __tablename__ = "false_state_stats"
 
     id = Column(INTEGER(), primary_key=True)
-    user_id = Column(INTEGER(), ForeignKey("users.id"))
+    user_id = Column(BIGINT(), ForeignKey("users.id"))
     played_at = Column(TIMESTAMP(timezone=True), default=datetime.now)
     won = Column(BOOLEAN())
     level = Column(INTEGER(), default=1)
