+++ backend/app/db/models/names_memory_stat.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column
-from sqlalchemy.dialects.postgresql import INTEGER, DATE
+from sqlalchemy.dialects.postgresql import INTEGER, DATE, BIGINT
 
 from backend.app.db import DeclarativeBase
 
@@ -7,7 +7,7 @@ class Names_Memory_Stat(DeclarativeBase):
     __tablename__ = 'names_memory_stats'
 
     id = Column(INTEGER(), primary_key=True, autoincrement=True)
-    user_id = Column(INTEGER(), nullable=False)
+    user_id = Column(BIGINT(), nullable=False)
     played_at = Column(DATE(), nullable=False)
     correct_answers = Column(INTEGER(), nullable=False)
     wrong_answers = Column(INTEGER(), nullable=False)
