+++ backend/app/db/models/viewed_image.py
@@ -1,5 +1,5 @@
 from sqlalchemy import Column
-from sqlalchemy.dialects.postgresql import INTEGER
+from sqlalchemy.dialects.postgresql import INTEGER, BIGINT
 
 from backend.app.db import DeclarativeBase
 
@@ -7,7 +7,7 @@ class Viewed_Image(DeclarativeBase):
     __tablename__ = 'viewed_images'
 
     id = Column(INTEGER(), primary_key=True, autoincrement=True)
-    user_id = Column(INTEGER(), nullable=False)
+    user_id = Column(BIGINT(), nullable=False)
     image_id = Column(INTEGER(), nullable=False)
     used_in_game = Column(INTEGER(), nullable=False)
     correct = Column(INTEGER(), nullable=False)
